[Refactor] Refactor the architecture of lab1
* Separate header and .c to differnet dirs.
* Separate terminal and commands in different file.
* Use a better way to implement commands for future extension.
