+++ lab1/Makefile
@@ -1,7 +1,11 @@
-SRCS = $(wildcard *.c)
-OBJS = $(SRCS:.c=.o)
-CFLAGS = -Wall -O3
-LDFLAGS = -nostdlib 
+SRC_DIR :=	src
+OBJ_DIR :=	obj
+
+SRCS = 		$(wildcard $(SRC_DIR)/*.c)
+OBJS = 		$(SRCS:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
+CPPFLAGS := 	-Iinclude -MMD -MP
+CFLAGS = 	-Wall -O3
+LDFLAGS = 	-nostdlib 
 
 CC = /home/elvis/Downloads/arm-gnu-toolchain-12.2.rel1-x86_64-aarch64-none-elf/bin/aarch64-none-elf-gcc
 LD = /home/elvis/Downloads/arm-gnu-toolchain-12.2.rel1-x86_64-aarch64-none-elf/bin/aarch64-none-elf-ld
@@ -14,8 +18,8 @@ all: clean kernel8.img
 start.o: start.S
 	$(CC) $(CFLAGS) -c -o $@ $<
 
-%.o: %.c
-	$(CC) $(CFLAGS) -c -o $@ $<
+$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c | $(OBJ_DIR)
+	$(CC) $(CPPFLAGS) $(CFLAGS) -c -o $@ $<
 
 kernel8.elf: start.o $(OBJS)
 	$(LD) $(LDFLAGS) $^ -T linker.ld -o $@
@@ -32,7 +36,11 @@ run_dbg: kernel8.img
 run: kernel8.img
 	$(QEMU) -M raspi3 -display none -serial null -serial stdio -kernel $<
 
+$(OBJ_DIR):
+	mkdir -p $@
+
 clean:
 	@rm kernel8.img *.o *.elf >/dev/null 2>/dev/null || true
+	@rm -r $(OBJ_DIR) >/dev/null 2>/dev/null || true
 
-	
+-include $(OBJS:.o=.d)
