+++ lab1/include/terminal.h
@@ -1,0 +1,10 @@
+#ifndef TERMINAL_H
+#define TERMINAL_H
+
+/// Main terminal
+void terminal_run(void);
+
+/// Call commands
+int command_exec(const char*);
+
+#endif //TERNINAL_H
