+++ lab1/src/str.c
